refactor(vc): extract vc_chain.go from vc_service.go
Move GetWorkflowVCChain and collectDIDResolutionBundle to a dedicated
chain-assembly file. vc_service.go is now a slim 159-line file with
the service struct, lifecycle methods, status summaries, and shared
helpers (hashData, generateVCID, marshalDataOrNull).

Pure file move: no symbol renames or logic changes.

Refs #415
